Cardano.TxSubmit.Util

Documentation

logException :: Trace IO TraceSubmitApi -> Text -> IO a -> IO a

ouroboros-network catches SomeException and if a nullTracer is passed into that code, the caught exception will not be logged. Therefore wrap all tx submission code that is called from network with an exception logger so at least the exception will be logged (instead of silently swallowed) and then rethrown.