Cardano.TxSubmit.Types

Documentation

type TxSubmitApi = "api" :> ToServantApi TxSubmitApiRecord

Servant API which provides access to tx submission webapi

newtype TxSubmitApiRecord route

A servant-generic record with all the methods of the API

Constructors

TxSubmitApiRecord

Fields

_txSubmitPost :: route :- ("submit" :> ("tx" :> (ReqBody '[CBORStream] ByteString :> PostAccepted '[JSON] TxId)))

data TxSubmitWebApiError

An error that can occur in the transaction submission web API.

Constructors

TxSubmitDecodeHex

TxSubmitEmpty

TxSubmitDecodeFail !DecoderError

TxSubmitBadTx !Text

TxSubmitFail TxCmdError

newtype TxSubmitPort

Constructors

TxSubmitPort Int

newtype EnvSocketError

Constructors

CliEnvVarLookup Text

data TxCmdError

Constructors

TxCmdSocketEnvError EnvSocketError

TxCmdTxReadError !RawCborDecodeError

TxCmdTxSubmitValidationError !TxValidationErrorInCardanoMode

newtype RawCborDecodeError

The errors that the raw CBOR transaction parsing/decoding functions can return.

Constructors

RawCborDecodeError [DecoderError]

renderTxCmdError :: TxCmdError -> Text